Crate hal_sys

Structs

HAL_CANStreamMessage

Storage for CAN Stream Messages.

HAL_ControlWord

HAL_JoystickAxes

HAL_JoystickButtons

HAL_JoystickDescriptor

HAL_JoystickPOVs

HAL_MatchInfo

HAL_Value

HAL Entry Value. Note this is a typed union.

__BindgenBitfieldUnit

Constants

HAL_AccelerometerRange_HAL_AccelerometerRange_k2G

HAL_AccelerometerRange_HAL_AccelerometerRange_k4G

HAL_AccelerometerRange_HAL_AccelerometerRange_k8G

HAL_AllianceStationID_HAL_AllianceStationID_kBlue1

HAL_AllianceStationID_HAL_AllianceStationID_kBlue2

HAL_AllianceStationID_HAL_AllianceStationID_kBlue3

HAL_AllianceStationID_HAL_AllianceStationID_kRed1

HAL_AllianceStationID_HAL_AllianceStationID_kRed2

HAL_AllianceStationID_HAL_AllianceStationID_kRed3

HAL_AllianceStationID_HAL_AllianceStationID_kUnknown

HAL_AnalogTriggerType_HAL_Trigger_kFallingPulse

HAL_AnalogTriggerType_HAL_Trigger_kInWindow

HAL_AnalogTriggerType_HAL_Trigger_kRisingPulse

HAL_AnalogTriggerType_HAL_Trigger_kState

HAL_CANDeviceType_HAL_CAN_Dev_kAccelerometer

Accelerometer.

HAL_CANDeviceType_HAL_CAN_Dev_kBroadcast

Broadcast.

HAL_CANDeviceType_HAL_CAN_Dev_kFirmwareUpdate

Firmware update.

HAL_CANDeviceType_HAL_CAN_Dev_kGearToothSensor

Gear tooth sensor.

HAL_CANDeviceType_HAL_CAN_Dev_kGyroSensor

Gyro sensor.

HAL_CANDeviceType_HAL_CAN_Dev_kIOBreakout

IO breakout.

HAL_CANDeviceType_HAL_CAN_Dev_kMiscellaneous

Miscellaneous.

HAL_CANDeviceType_HAL_CAN_Dev_kMotorController

Motor controller.

HAL_CANDeviceType_HAL_CAN_Dev_kPneumatics

Pneumatics.

HAL_CANDeviceType_HAL_CAN_Dev_kPowerDistribution

Power distribution.

HAL_CANDeviceType_HAL_CAN_Dev_kRelayController

Relay controller.

HAL_CANDeviceType_HAL_CAN_Dev_kRobotController

Robot controller.

HAL_CANDeviceType_HAL_CAN_Dev_kUltrasonicSensor

Ultrasonic sensor.

HAL_CANManufacturer_HAL_CAN_Man_kAndyMark

AndyMark.

HAL_CANManufacturer_HAL_CAN_Man_kBroadcast

Broadcast.

HAL_CANManufacturer_HAL_CAN_Man_kCTRE

Cross the Road Electronics.

HAL_CANManufacturer_HAL_CAN_Man_kCopperforge

Copperforge.

HAL_CANManufacturer_HAL_CAN_Man_kDEKA

DEKA Research and Development Corp.

HAL_CANManufacturer_HAL_CAN_Man_kGrapple

Grapple.

HAL_CANManufacturer_HAL_CAN_Man_kKauaiLabs

Kauai Labs.

HAL_CANManufacturer_HAL_CAN_Man_kLM

Luminary Micro.

HAL_CANManufacturer_HAL_CAN_Man_kMS

MindSensors.

HAL_CANManufacturer_HAL_CAN_Man_kNI

National Instruments.

HAL_CANManufacturer_HAL_CAN_Man_kPWF

Playing With Fusion.

HAL_CANManufacturer_HAL_CAN_Man_kREV

REV robotics.

HAL_CANManufacturer_HAL_CAN_Man_kReduxRobotics

Redux Robotics.

HAL_CANManufacturer_HAL_CAN_Man_kStudica

Studica.

HAL_CANManufacturer_HAL_CAN_Man_kTeamUse

Team use.

HAL_CANManufacturer_HAL_CAN_Man_kTheThriftyBot

TheThriftyBot.

HAL_CANManufacturer_HAL_CAN_Man_kVividHosting

Vivid-Hosting.

HAL_CAN_BUFFER_OVERRUN

HAL_CAN_BUFFER_OVERRUN_MESSAGE

HAL_CAN_IS_FRAME_11BIT

HAL_CAN_IS_FRAME_REMOTE

HAL_CAN_SEND_PERIOD_NO_REPEAT

HAL_CAN_SEND_PERIOD_STOP_REPEATING

HAL_CAN_TIMEOUT

HAL_CAN_TIMEOUT_MESSAGE

HAL_CONSOLE_OUT_ENABLED_ERROR

HAL_CONSOLE_OUT_ENABLED_ERROR_MESSAGE

HAL_COUNTER_NOT_SUPPORTED

HAL_COUNTER_NOT_SUPPORTED_MESSAGE

HAL_Counter_Mode_HAL_Counter_kExternalDirection

HAL_Counter_Mode_HAL_Counter_kPulseLength

HAL_Counter_Mode_HAL_Counter_kSemiperiod

HAL_Counter_Mode_HAL_Counter_kTwoPulse

HAL_ERR_CANSessionMux_InvalidBuffer

HAL_ERR_CANSessionMux_MessageNotFound

HAL_ERR_CANSessionMux_NotAllowed

HAL_ERR_CANSessionMux_NotInitialized

HAL_ERR_CANSessionMux_SessionOverrun

HAL_EncoderEncodingType_HAL_Encoder_k1X

HAL_EncoderEncodingType_HAL_Encoder_k2X

HAL_EncoderEncodingType_HAL_Encoder_k4X

HAL_EncoderIndexingType_HAL_kResetOnFallingEdge

HAL_EncoderIndexingType_HAL_kResetOnRisingEdge

HAL_EncoderIndexingType_HAL_kResetWhileHigh

HAL_EncoderIndexingType_HAL_kResetWhileLow

HAL_HANDLE_ERROR

HAL_HANDLE_ERROR_MESSAGE

HAL_I2CPort_HAL_I2C_kInvalid

HAL_I2CPort_HAL_I2C_kMXP

HAL_I2CPort_HAL_I2C_kOnboard

HAL_INVALID_ACCUMULATOR_CHANNEL

HAL_INVALID_ACCUMULATOR_CHANNEL_MESSAGE

HAL_INVALID_DMA_ADDITION

HAL_INVALID_DMA_ADDITION_MESSAGE

HAL_INVALID_DMA_STATE

HAL_INVALID_DMA_STATE_MESSAGE

HAL_IO_CONFIG_DATA_SIZE

HAL_LED_CHANNEL_ERROR

HAL_LED_CHANNEL_ERROR_MESSAGE

HAL_MatchType_HAL_kMatchType_elimination

HAL_MatchType_HAL_kMatchType_none

HAL_MatchType_HAL_kMatchType_practice

HAL_MatchType_HAL_kMatchType_qualification

HAL_PWM_SCALE_ERROR

HAL_PWM_SCALE_ERROR_MESSAGE

HAL_RadioLEDState_HAL_RadioLED_kGreen

HAL_RadioLEDState_HAL_RadioLED_kOff

HAL_RadioLEDState_HAL_RadioLED_kOrange

HAL_RadioLEDState_HAL_RadioLED_kRed

HAL_RuntimeType_HAL_Runtime_RoboRIO

HAL_RuntimeType_HAL_Runtime_RoboRIO2

HAL_RuntimeType_HAL_Runtime_Simulation

HAL_SERIAL_PORT_ERROR

HAL_SERIAL_PORT_ERROR_MESSAGE

HAL_SERIAL_PORT_NOT_FOUND

HAL_SERIAL_PORT_NOT_FOUND_MESSAGE

HAL_SERIAL_PORT_OPEN_ERROR

HAL_SERIAL_PORT_OPEN_ERROR_MESSAGE

HAL_SIM_NOT_SUPPORTED

HAL_SIM_NOT_SUPPORTED_MESSAGE

HAL_SPIMode_HAL_SPI_kMode0

HAL_SPIMode_HAL_SPI_kMode1

HAL_SPIMode_HAL_SPI_kMode2

HAL_SPIMode_HAL_SPI_kMode3

HAL_SPIPort_HAL_SPI_kInvalid

HAL_SPIPort_HAL_SPI_kMXP

HAL_SPIPort_HAL_SPI_kOnboardCS0

HAL_SPIPort_HAL_SPI_kOnboardCS1

HAL_SPIPort_HAL_SPI_kOnboardCS2

HAL_SPIPort_HAL_SPI_kOnboardCS3

HAL_SUCCESS

HAL_SYS_STATUS_DATA_SIZE

HAL_SerialPort_HAL_SerialPort_MXP

HAL_SerialPort_HAL_SerialPort_Onboard

HAL_SerialPort_HAL_SerialPort_USB1

HAL_SerialPort_HAL_SerialPort_USB2

HAL_SimValueDirection_HAL_SimValueBidir

< bidirectional between user code and simulator

HAL_SimValueDirection_HAL_SimValueInput

< input to user code from the simulator

HAL_SimValueDirection_HAL_SimValueOutput

< output from user code to the simulator

HAL_THREAD_PRIORITY_ERROR

HAL_THREAD_PRIORITY_ERROR_MESSAGE

HAL_THREAD_PRIORITY_RANGE_ERROR

HAL_THREAD_PRIORITY_RANGE_ERROR_MESSAGE

HAL_Type_HAL_BOOLEAN

HAL_Type_HAL_DOUBLE

HAL_Type_HAL_ENUM

HAL_Type_HAL_INT

HAL_Type_HAL_LONG

HAL_Type_HAL_UNASSIGNED

HAL_USER_STATUS_DATA_SIZE

HAL_USE_LAST_ERROR

HAL_USE_LAST_ERROR_MESSAGE

HAL_WARN_CANSessionMux_NoToken

HAL_kInvalidHandle

HAL_kMaxJoystickAxes

HAL_kMaxJoystickPOVs

HAL_kMaxJoysticks

WPI_VERSION

Functions

HAL_AllocateDigitalPWM⚠

Allocates a DO PWM Generator.

HAL_CAN_CloseStreamSession⚠

Closes a CAN stream.

HAL_CAN_GetCANStatus⚠

Gets CAN status information.

HAL_CAN_OpenStreamSession⚠

Opens a CAN stream.

HAL_CAN_ReadStreamSession⚠

Reads a CAN stream message.

HAL_CAN_ReceiveMessage⚠

Receives a CAN message.

HAL_CAN_SendMessage⚠

Sends a CAN message.

HAL_CalibrateAnalogGyro⚠

Calibrates the analog gyro.

HAL_CancelNotifierAlarm⚠

Cancels the next notifier alarm.

HAL_CheckAnalogInputChannel⚠

Checks that the analog output channel number is valid. Verifies that the analog channel number is one of the legal channel numbers. Channel numbers are 0-based.

HAL_CheckAnalogModule⚠

Checks that the analog module number is valid.

HAL_CheckAnalogOutputChannel⚠

Checks that the analog output channel number is valid.

HAL_CheckCTREPCMSolenoidChannel⚠

Checks if a solenoid channel number is valid.

HAL_CheckDIOChannel⚠

Checks if a DIO channel is valid.

HAL_CheckPWMChannel⚠

Checks if a pwm channel is valid.

HAL_CheckRelayChannel⚠

Checks if a relay channel is valid.

HAL_CleanAnalogTrigger⚠

Frees an analog trigger.

HAL_CleanCAN⚠

Frees a CAN device

HAL_CleanInterrupts⚠

Frees an interrupt.

HAL_CleanNotifier⚠

Cleans a notifier.

HAL_ClearAllCTREPCMStickyFaults⚠

Clears all sticky faults on this device.

HAL_ClearCounterDownSource⚠

Disables the down counting source to the counter.

HAL_ClearCounterUpSource⚠

Disables the up counting source to the counter.

HAL_ClearSerial⚠

Clears the receive buffer of the serial port.

HAL_CloseI2C⚠

Closes an I2C port

HAL_CloseSPI⚠

Closes the SPI port.

HAL_CloseSerial⚠

Closes a serial port.

HAL_ConfigureSPIAutoStall⚠

Configure the Auto SPI Stall time between reads.

HAL_CreateSimDevice⚠

Creates a simulated device.

HAL_CreateSimValue⚠

Creates a value on a simulated device.

HAL_CreateSimValueEnum⚠

Creates an enumerated value on a simulated device.

HAL_CreateSimValueEnumDouble⚠

Creates an enumerated value on a simulated device with double values.

HAL_DisableSerialTermination⚠

Disables a termination character for reads.

HAL_EnableSerialTermination⚠

Sets the termination character that terminates a read.

HAL_ExitMain⚠

Causes HAL_RunMain() to exit.

HAL_ExpandFPGATime⚠

Given an 32 bit FPGA time, expand it to the nearest likely 64 bit FPGA time.

HAL_FireCTREPCMOneShot⚠

Fire a single solenoid shot.

HAL_FlushSerial⚠

Flushes the serial write buffer out to the port.

HAL_ForceSPIAutoRead⚠

Immediately forces an SPI read to happen.

HAL_FreeAnalogGyro⚠

Frees an analog gyro.

HAL_FreeAnalogInputPort⚠

Frees an analog input port.

HAL_FreeAnalogOutputPort⚠

Frees an analog output port.

HAL_FreeCTREPCM⚠

Frees a PCM handle.

HAL_FreeCounter⚠

Frees a counter.

HAL_FreeDIOPort⚠

Frees a DIO port.

HAL_FreeDigitalPWM⚠

Frees the resource associated with a DO PWM generator.

HAL_FreeEncoder⚠

Frees an encoder.

HAL_FreeJoystickName⚠

Frees a joystick name received with HAL_GetJoystickName

HAL_FreePWMPort⚠

Frees a PWM port.

HAL_FreeRelayPort⚠

Frees a relay port.

HAL_FreeSPIAuto⚠

Frees an SPI automatic accumulator.

HAL_FreeSimDevice⚠

Frees a simulated device.

HAL_GetAccelerometerX⚠

Gets the x-axis acceleration.

HAL_GetAccelerometerY⚠

Gets the y-axis acceleration.

HAL_GetAccelerometerZ⚠

Gets the z-axis acceleration.

HAL_GetAccumulatorCount⚠

Read the number of accumulated values.

HAL_GetAccumulatorOutput⚠

Read the accumulated value and the number of accumulated values atomically.

HAL_GetAccumulatorValue⚠

Read the accumulated value.

HAL_GetAllJoystickData⚠

HAL_GetAllianceStation⚠

Gets the current alliance station ID.

HAL_GetAnalogAverageBits⚠

Gets the number of averaging bits.

HAL_GetAnalogAverageValue⚠

Gets a sample from the output of the oversample and average engine for the channel.

HAL_GetAnalogAverageVoltage⚠

Gets a scaled sample from the output of the oversample and average engine for the channel.

HAL_GetAnalogGyroAngle⚠

Gets the gyro angle in degrees.

HAL_GetAnalogGyroCenter⚠

Gets the calibrated gyro center.

HAL_GetAnalogGyroOffset⚠

Gets the calibrated gyro offset.

HAL_GetAnalogGyroRate⚠

Gets the gyro rate in degrees/second.

HAL_GetAnalogLSBWeight⚠

Gets the factory scaling least significant bit weight constant. The least significant bit weight constant for the channel that was calibrated in manufacturing and stored in an eeprom in the module.

HAL_GetAnalogOffset⚠

Gets the factory scaling offset constant. The offset constant for the channel that was calibrated in manufacturing and stored in an eeprom in the module.

HAL_GetAnalogOutput⚠

Gets the current analog output value.

HAL_GetAnalogOversampleBits⚠

Gets the number of oversample bits.

HAL_GetAnalogSampleRate⚠

Gets the current sample rate.

HAL_GetAnalogTriggerFPGAIndex⚠

Get the FPGA index for the AnlogTrigger.

HAL_GetAnalogTriggerInWindow⚠

Returns the InWindow output of the analog trigger.

HAL_GetAnalogTriggerOutput⚠

Gets the state of the analog trigger output.

HAL_GetAnalogTriggerTriggerState⚠

Returns the TriggerState output of the analog trigger.

HAL_GetAnalogValue⚠

Gets a sample straight from the channel on this module.

HAL_GetAnalogValueToVolts⚠

Get the analog voltage from a raw value.

HAL_GetAnalogVoltage⚠

Gets a scaled sample straight from the channel on this module.

HAL_GetAnalogVoltsToValue⚠

Converts a voltage to a raw value for a specified channel.

HAL_GetBrownedOut⚠

Gets if the system is in a browned out state.

HAL_GetBrownoutVoltage⚠

Get the current brownout voltage setting.

HAL_GetCANPacketBaseTime⚠

Reads the current value of the millisecond-resolution timer that the CAN API functions use as a time base.

HAL_GetCPUTemp⚠

Get the current CPU temperature in degrees Celsius

HAL_GetCTREPCMClosedLoopControl⚠

Get whether the PCM closed loop control is enabled.

HAL_GetCTREPCMCompressor⚠

Get whether compressor is turned on.

HAL_GetCTREPCMCompressorCurrent⚠

Returns the current drawn by the compressor.

HAL_GetCTREPCMCompressorCurrentTooHighFault⚠

Return whether the compressor current is currently too high.

HAL_GetCTREPCMCompressorCurrentTooHighStickyFault⚠

Returns whether the compressor current has been too high since sticky faults were last cleared. This fault is persistent and can be cleared by HAL_ClearAllCTREPCMStickyFaults()

HAL_GetCTREPCMCompressorNotConnectedFault⚠

Returns whether the compressor is currently disconnected.

HAL_GetCTREPCMCompressorNotConnectedStickyFault⚠

Returns whether the compressor has been disconnected since sticky faults were last cleared. This fault is persistent and can be cleared by HAL_ClearAllCTREPCMStickyFaults()

HAL_GetCTREPCMCompressorShortedFault⚠

Returns whether the compressor is currently shorted.

HAL_GetCTREPCMCompressorShortedStickyFault⚠

Returns whether the compressor has been shorted since sticky faults were last cleared. This fault is persistent and can be cleared by HAL_ClearAllCTREPCMStickyFaults()

HAL_GetCTREPCMPressureSwitch⚠

Returns the state of the pressure switch.

HAL_GetCTREPCMSolenoidDisabledList⚠

Get a bitmask of disabled solenoids.

HAL_GetCTREPCMSolenoidVoltageFault⚠

Returns whether the solenoid is currently reporting a voltage fault.

HAL_GetCTREPCMSolenoidVoltageStickyFault⚠

Returns whether the solenoid has reported a voltage fault since sticky faults were last cleared. This fault is persistent and can be cleared by HAL_ClearAllCTREPCMStickyFaults()

HAL_GetCTREPCMSolenoids⚠

Gets a bitmask of solenoid values.

HAL_GetComments⚠

Returns the comments from the roboRIO web interface.

HAL_GetControlWord⚠

Gets the current control word of the driver station.

HAL_GetCounter⚠

Reads the current counter value.

HAL_GetCounterDirection⚠

Gets the last direction the counter value changed.

HAL_GetCounterPeriod⚠

Gets the Period of the most recent count.

HAL_GetCounterSamplesToAverage⚠

Gets the Samples to Average which specifies the number of samples of the timer to average when calculating the period. Perform averaging to account for mechanical imperfections or as oversampling to increase resolution.

HAL_GetCounterStopped⚠

Determines if the clock is stopped.

HAL_GetCurrentThreadPriority⚠

Gets the thread priority for the current thread.

HAL_GetDIO⚠

Reads a digital value from a DIO channel.

HAL_GetDIODirection⚠

Reads the direction of a DIO channel.

HAL_GetEncoder⚠

Gets the current counts of the encoder after encoding type scaling.

HAL_GetEncoderDecodingScaleFactor⚠

Gets the decoding scale factor of the encoder.

HAL_GetEncoderDirection⚠

Gets the last direction the encoder value changed.

HAL_GetEncoderDistance⚠

Gets the current distance traveled by the encoder.

HAL_GetEncoderDistancePerPulse⚠

Gets the user set distance per pulse of the encoder.

HAL_GetEncoderEncodingScale⚠

Gets the encoder scale value.

HAL_GetEncoderEncodingType⚠

Gets the encoding type of the encoder.

HAL_GetEncoderFPGAIndex⚠

Gets the FPGA index of the encoder.

HAL_GetEncoderPeriod⚠

Gets the Period of the most recent count.

HAL_GetEncoderRate⚠

Gets the current rate of the encoder.

HAL_GetEncoderRaw⚠

Gets the raw counts of the encoder.

HAL_GetEncoderSamplesToAverage⚠

Gets the current samples to average value.

HAL_GetEncoderStopped⚠

Determines if the clock is stopped.

HAL_GetErrorMessage⚠

Gets the error message for a specific status code.

HAL_GetFPGAButton⚠

Gets the state of the “USER” button on the roboRIO.

HAL_GetFPGARevision⚠

Returns the FPGA Revision number.

HAL_GetFPGATime⚠

Reads the microsecond-resolution timer on the FPGA.

HAL_GetFPGAVersion⚠

Returns the FPGA Version number.

HAL_GetFilterPeriod⚠

Gets the filter period for the specified filter index.

HAL_GetFilterSelect⚠

Reads the filter index from the FPGA.

HAL_GetJoystickAxes⚠

Gets the axes of a specific joystick.

HAL_GetJoystickAxisType⚠

Gets the type of a specific joystick axis.

HAL_GetJoystickButtons⚠

Gets the buttons of a specific joystick.

HAL_GetJoystickDescriptor⚠

Retrieves the Joystick Descriptor for particular slot.

HAL_GetJoystickIsXbox⚠

Gets whether a specific joystick is considered to be an XBox controller.

HAL_GetJoystickName⚠

Gets the name of a joystick.

HAL_GetJoystickPOVs⚠

Gets the POVs of a specific joystick.

HAL_GetJoystickType⚠

Gets the type of joystick connected.

HAL_GetLastError⚠

Gets the last error set on this thread, or the message for the status code.

HAL_GetMatchInfo⚠

Gets info about a specific match.

HAL_GetMatchTime⚠

Return the approximate match time. The FMS does not send an official match time to the robots, but does send an approximate match time. The value will count down the time remaining in the current period (auto or teleop). Warning: This is not an official time (so it cannot be used to dispute ref calls or guarantee that a function will trigger before the match ends).

HAL_GetNumAccumulators⚠

Gets the number of analog accumulators in the current system.

HAL_GetNumAddressableLEDs⚠

Gets the number of addressable LED generators in the current system.

HAL_GetNumAnalogInputs⚠

Gets the number of analog inputs in the current system.

HAL_GetNumAnalogOutputs⚠

Gets the number of analog outputs in the current system.

HAL_GetNumAnalogTriggers⚠

Gets the number of analog triggers in the current system.

HAL_GetNumCTREPCMModules⚠

Gets the number of PCM modules in the current system.

HAL_GetNumCTREPDPChannels⚠

Gets the number of PDP channels in the current system.

HAL_GetNumCTREPDPModules⚠

Gets the number of PDP modules in the current system.

HAL_GetNumCTRESolenoidChannels⚠

Gets the number of solenoid channels in the current system.

HAL_GetNumCounters⚠

Gets the number of counters in the current system.

HAL_GetNumDigitalChannels⚠

Gets the number of digital channels in the current system.

HAL_GetNumDigitalHeaders⚠

Gets the number of digital headers in the current system.

HAL_GetNumDigitalPWMOutputs⚠

Gets the number of digital IO PWM outputs in the current system.

HAL_GetNumDutyCycles⚠

Gets the number of duty cycle inputs in the current system.

HAL_GetNumEncoders⚠

Gets the number of quadrature encoders in the current system.

HAL_GetNumInterrupts⚠

Gets the number of interrupts in the current system.

HAL_GetNumPWMChannels⚠

Gets the number of PWM channels in the current system.

HAL_GetNumPWMHeaders⚠

Gets the number of PWM headers in the current system.

HAL_GetNumREVPDHChannels⚠

Gets the number of PDH channels in the current system.

HAL_GetNumREVPDHModules⚠

Gets the number of PDH modules in the current system.

HAL_GetNumREVPHChannels⚠

Gets the number of PH channels in the current system.

HAL_GetNumREVPHModules⚠

Gets the number of PH modules in the current system.

HAL_GetNumRelayChannels⚠

Gets the number of relay channels in the current system.

HAL_GetNumRelayHeaders⚠

Gets the number of relay headers in the current system.

HAL_GetOutputsEnabled⚠

Gets if outputs are enabled by the control system.

HAL_GetPWMConfigMicroseconds⚠

Gets the pwm configuration settings for the PWM channel.

HAL_GetPWMCycleStartTime⚠

Gets the pwm starting cycle time.

HAL_GetPWMEliminateDeadband⚠

Gets the current eliminate deadband value.

HAL_GetPWMLoopTiming⚠

Gets the loop timing of the PWM system.

HAL_GetPWMPosition⚠

Gets a position value from a PWM channel.

HAL_GetPWMPulseTimeMicroseconds⚠

Gets the current microsecond pulse time from a PWM channel.

HAL_GetPWMSpeed⚠

Gets a scaled value from a PWM channel.

HAL_GetPort⚠

Gets a port handle for a specific channel.

HAL_GetPortWithModule⚠

Gets a port handle for a specific channel and module.

HAL_GetRSLState⚠

Gets the current state of the Robot Signal Light (RSL).

HAL_GetRadioLEDState⚠

Get the state of the “Radio” LED.

HAL_GetRelay⚠

Gets the current state of the relay channel.

HAL_GetRuntimeType⚠

Returns the runtime type of the HAL.

HAL_GetSPIAutoDroppedCount⚠

Gets the count of how many SPI accumulations have been missed.

HAL_GetSPIHandle⚠

Gets the stored handle for a SPI port.

HAL_GetSPIMode⚠

Gets the SPI Mode.

HAL_GetSerialBytesReceived⚠

Gets the number of bytes currently in the read buffer.

HAL_GetSerialFD⚠

Gets the raw serial port file descriptor from a handle.

HAL_GetSerialNumber⚠

Returns the roboRIO serial number.

HAL_GetSimDeviceName⚠

Get the name of a simulated device

HAL_GetSimValue⚠

Gets a simulated value.

HAL_GetSystemActive⚠

Gets if the system outputs are currently active.

HAL_GetSystemClockTicksPerMicrosecond⚠

Gets the number of FPGA system clock ticks per microsecond.

HAL_GetSystemTimeValid⚠

Gets if the system time is valid.

HAL_GetTeamNumber⚠

Returns the team number configured for the robot controller. @return team number, or 0 if not found.

HAL_GetThreadPriority⚠

Gets the thread priority for the specified thread.

HAL_GetUserActive3V3⚠

Gets the active state of the 3V3 rail.

HAL_GetUserActive5V⚠

Gets the active state of the 5V rail.

HAL_GetUserActive6V⚠

Gets the active state of the 6V rail.

HAL_GetUserCurrent3V3⚠

Gets the 3V3 rail current.

HAL_GetUserCurrent5V⚠

Gets the 5V rail current.

HAL_GetUserCurrent6V⚠

Gets the 6V rail current.

HAL_GetUserCurrentFaults3V3⚠

Gets the fault count for the 3V3 rail.

HAL_GetUserCurrentFaults5V⚠

Gets the fault count for the 5V rail.

HAL_GetUserCurrentFaults6V⚠

Gets the fault count for the 6V rail.

HAL_GetUserVoltage3V3⚠

Gets the 3V3 rail voltage.

HAL_GetUserVoltage5V⚠

Gets the 5V rail voltage.

HAL_GetUserVoltage6V⚠

Gets the 6V rail voltage.

HAL_GetVinCurrent⚠

Gets the roboRIO input current.

HAL_GetVinVoltage⚠

Gets the roboRIO input voltage.

HAL_HasMain⚠

Returns true if HAL_SetMain() has been called.

HAL_InitAccumulator⚠

Initialize the accumulator.

HAL_InitSPIAuto⚠

Initializes the SPI automatic accumulator.

HAL_Initialize⚠

Call this to start up HAL. This is required for robot programs.

HAL_InitializeAnalogGyro⚠

Initializes an analog gyro.

HAL_InitializeAnalogInputPort⚠

Initializes the analog input port using the given port object.

HAL_InitializeAnalogOutputPort⚠

Initializes the analog output port using the given port object.

HAL_InitializeAnalogTrigger⚠

Initializes an analog trigger.

HAL_InitializeAnalogTriggerDutyCycle⚠

Initializes an analog trigger with a Duty Cycle input

HAL_InitializeCAN⚠

Initializes a CAN device.

HAL_InitializeCTREPCM⚠

Initializes a PCM.

HAL_InitializeCounter⚠

Initializes a counter.

HAL_InitializeDIOPort⚠

Creates a new instance of a digital port.

HAL_InitializeEncoder⚠

Initializes an encoder.

HAL_InitializeI2C⚠

Initializes the I2C port.

HAL_InitializeInterrupts⚠

Initializes an interrupt.

HAL_InitializeNotifier⚠

Initializes a notifier.

HAL_InitializePWMPort⚠

Initializes a PWM port.

HAL_InitializeRelayPort⚠

Initializes a relay.

HAL_InitializeSPI⚠

Initializes the SPI port. Opens the port if necessary and saves the handle.

HAL_InitializeSerialPort⚠

Initializes a serial port.

HAL_InitializeSerialPortDirect⚠

Initializes a serial port with a direct name.

HAL_IsAccumulatorChannel⚠

Is the channel attached to an accumulator.

HAL_IsAnyPulsing⚠

Checks if any DIO line is currently generating a pulse.

HAL_IsPulsing⚠

Checks a DIO line to see if it is currently generating a pulse.

HAL_LatchPWMZero⚠

Forces a PWM signal to go to 0 temporarily.

HAL_ObserveUserProgramAutonomous⚠

Sets the autonomous enabled flag in the DS.

HAL_ObserveUserProgramDisabled⚠

Sets the disabled flag in the DS.

HAL_ObserveUserProgramStarting⚠

Sets the program starting flag in the DS.

HAL_ObserveUserProgramTeleop⚠

Sets the teleoperated enabled flag in the DS.

HAL_ObserveUserProgramTest⚠

Sets the test mode flag in the DS.

HAL_ProvideNewDataEventHandle⚠

HAL_Pulse⚠

Generates a single digital pulse.

HAL_PulseMultiple⚠

Generates a single digital pulse on multiple channels.

HAL_ReadCANPacketLatest⚠

Reads a CAN packet. The will continuously return the last packet received, without accounting for packet age.

HAL_ReadCANPacketNew⚠

Reads a new CAN packet.

HAL_ReadCANPacketTimeout⚠

Reads a CAN packet. The will return the last packet received until the packet is older then the requested timeout. Then it will return an error code.

HAL_ReadI2C⚠

Executes a read transaction with the device.

HAL_ReadInterruptFallingTimestamp⚠

Returns the timestamp for the falling interrupt that occurred most recently.

HAL_ReadInterruptRisingTimestamp⚠

Returns the timestamp for the rising interrupt that occurred most recently.

HAL_ReadSPI⚠

Executes a read from the device.

HAL_ReadSPIAutoReceivedData⚠

Reads data received by the SPI accumulator. Each received data sequence consists of a timestamp followed by the received data bytes, one byte per word (in the least significant byte). The length of each received data sequence is the same as the combined dataSize + zeroSize set in HAL_SetSPIAutoTransmitData.

HAL_ReadSerial⚠

Reads data from the serial port.

HAL_RefreshDSData⚠

Refresh the DS control word.

HAL_ReleaseWaitingInterrupt⚠

Releases a waiting interrupt.

HAL_RemoveNewDataEventHandle⚠

HAL_Report⚠

Reports a hardware usage to the HAL.

HAL_RequestInterrupts⚠

Requests interrupts on a specific digital source.

HAL_ResetAccumulator⚠

Resets the accumulator to the initial value.

HAL_ResetAnalogGyro⚠

Resets the analog gyro value to 0.

HAL_ResetCounter⚠

Resets the Counter to zero.

HAL_ResetEncoder⚠

Reads the current encoder value.

HAL_ResetSimValue⚠

Resets a simulated double or integral value to 0. Has no effect on other value types. Use this instead of Set(0) for resetting incremental sensor values like encoder counts or gyro accumulated angle to ensure correct behavior in a distributed system (e.g. WebSockets).

HAL_RunMain⚠

Runs the main function provided to HAL_SetMain().

HAL_SendConsoleLine⚠

Sends a line to the driver station console.

HAL_SendError⚠

Sends an error to the driver station.

HAL_SetAccelerometerActive⚠

Sets the accelerometer to active or standby mode.

HAL_SetAccelerometerRange⚠

Sets the range of values that can be measured (either 2, 4, or 8 g-forces).

HAL_SetAccumulatorCenter⚠

Set the center value of the accumulator.

HAL_SetAccumulatorDeadband⚠

Set the accumulator’s deadband.

HAL_SetAnalogAverageBits⚠

Sets the number of averaging bits.

HAL_SetAnalogGyroDeadband⚠

Sets the deadband of the analog gyro.

HAL_SetAnalogGyroParameters⚠

Sets the analog gyro parameters to the specified values.

HAL_SetAnalogGyroVoltsPerDegreePerSecond⚠

Sets the analog gyro volts per degrees per second scaling.

HAL_SetAnalogInputSimDevice⚠

Indicates the analog input is used by a simulated device.

HAL_SetAnalogOutput⚠

Sets an analog output value.

HAL_SetAnalogOversampleBits⚠

Sets the number of oversample bits.

HAL_SetAnalogSampleRate⚠

Sets the sample rate.

HAL_SetAnalogTriggerAveraged⚠

Configures the analog trigger to use the averaged vs. raw values.

HAL_SetAnalogTriggerFiltered⚠

Configures the analog trigger to use a filtered value.

HAL_SetAnalogTriggerLimitsDutyCycle⚠

Sets the upper and lower limits of the analog trigger.

HAL_SetAnalogTriggerLimitsRaw⚠

Sets the raw ADC upper and lower limits of the analog trigger.

HAL_SetAnalogTriggerLimitsVoltage⚠

Sets the upper and lower limits of the analog trigger.

HAL_SetBrownoutVoltage⚠

Set the voltage the roboRIO will brownout and disable all outputs.

HAL_SetCTREPCMClosedLoopControl⚠

Enables the compressor closed loop control using the digital pressure switch. The compressor will turn on when the pressure switch indicates that the system is not full, and will turn off when the pressure switch indicates that the system is full.

HAL_SetCTREPCMOneShotDuration⚠

Set the duration for a single solenoid shot.

HAL_SetCTREPCMSolenoids⚠

Sets solenoids on a pneumatics module.

HAL_SetCounterAverageSize⚠

Sets the average sample size of a counter.

HAL_SetCounterDownSource⚠

Sets the source object that causes the counter to count down.

HAL_SetCounterDownSourceEdge⚠

Sets the down source to either detect rising edges or falling edges. Note that both are allowed to be set true at the same time without issues.

HAL_SetCounterExternalDirectionMode⚠

Sets directional counting mode on this counter.

HAL_SetCounterMaxPeriod⚠

Sets the maximum period where the device is still considered “moving”.

HAL_SetCounterPulseLengthMode⚠

Configures the counter to count in up or down based on the length of the input pulse.

HAL_SetCounterReverseDirection⚠

Sets the Counter to return reversed sensing on the direction.

HAL_SetCounterSamplesToAverage⚠

Sets the Samples to Average which specifies the number of samples of the timer to average when calculating the period. Perform averaging to account for mechanical imperfections or as oversampling to increase resolution.

HAL_SetCounterSemiPeriodMode⚠

Sets Semi-period mode on this counter.

HAL_SetCounterUpDownMode⚠

Sets standard up / down counting mode on this counter.

HAL_SetCounterUpSource⚠

Sets the source object that causes the counter to count up.

HAL_SetCounterUpSourceEdge⚠

Sets the up source to either detect rising edges or falling edges.

HAL_SetCounterUpdateWhenEmpty⚠

Selects whether you want to continue updating the event timer output when there are no samples captured.

HAL_SetCurrentThreadPriority⚠

Sets the thread priority for the current thread.

HAL_SetDIO⚠

Writes a digital value to a DIO channel.

HAL_SetDIODirection⚠

Sets the direction of a DIO channel.

HAL_SetDIOSimDevice⚠

Indicates the DIO channel is used by a simulated device.

HAL_SetDigitalPWMDutyCycle⚠

Configures the duty-cycle of the PWM generator.

HAL_SetDigitalPWMOutputChannel⚠

Configures which DO channel the PWM signal is output on.

HAL_SetDigitalPWMPPS⚠

Configures the digital PWM to be a PPS signal with specified duty cycle.

HAL_SetDigitalPWMRate⚠

Changes the frequency of the DO PWM generator.

HAL_SetEncoderDistancePerPulse⚠

Sets the distance traveled per encoder pulse. This is used as a scaling factor for the rate and distance calls.

HAL_SetEncoderIndexSource⚠

Sets the source for an index pulse on the encoder.

HAL_SetEncoderMaxPeriod⚠

Sets the maximum period where the device is still considered “moving”.

HAL_SetEncoderMinRate⚠

Sets the minimum rate to be considered moving by the encoder.

HAL_SetEncoderReverseDirection⚠

Sets if to reverse the direction of the encoder.

HAL_SetEncoderSamplesToAverage⚠

Sets the number of encoder samples to average when calculating encoder rate.

HAL_SetEncoderSimDevice⚠

Indicates the encoder is used by a simulated device.

HAL_SetFilterPeriod⚠

Sets the filter period for the specified filter index.

HAL_SetFilterSelect⚠

Writes the filter index from the FPGA.

HAL_SetInterruptUpSourceEdge⚠

Sets the edges to trigger the interrupt on.

HAL_SetJoystickOutputs⚠

Set joystick outputs.

HAL_SetMain⚠

Sets up the system to run the provided main loop in the main thread (e.g. the thread in which main() starts execution) and run the robot code in a separate thread.

HAL_SetNotifierName⚠

Sets the name of a notifier.

HAL_SetNotifierThreadPriority⚠

Sets the HAL notifier thread priority.

HAL_SetPWMAlwaysHighMode⚠

Sets the PWM output to be a continuous high signal while enabled.

HAL_SetPWMConfigMicroseconds⚠

Sets the configuration settings for the PWM channel.

HAL_SetPWMDisabled⚠

Sets a PWM channel to be disabled.

HAL_SetPWMEliminateDeadband⚠

Sets if the FPGA should output the center value if the input value is within the deadband.

HAL_SetPWMPeriodScale⚠

Sets how how often the PWM signal is squelched, thus scaling the period.

HAL_SetPWMPosition⚠

Sets a PWM channel to the desired position value.

HAL_SetPWMPulseTimeMicroseconds⚠

Sets a PWM channel to the desired pulse width in microseconds.

HAL_SetPWMSpeed⚠

Sets a PWM channel to the desired scaled value.

HAL_SetPrintErrorImpl⚠

Set the print function used by HAL_SendError

HAL_SetRadioLEDState⚠

Set the state of the “Radio” LED. @param state The state to set the LED to. @param[out] status the error code, or 0 for success

HAL_SetRelay⚠

Sets the state of a relay output.

HAL_SetSPIAutoTransmitData⚠

Sets the data to be transmitted to the device to initiate a read.

HAL_SetSPIChipSelectActiveHigh⚠

Sets the CS Active high for a SPI port.

HAL_SetSPIChipSelectActiveLow⚠

Sets the CS Active low for a SPI port.

HAL_SetSPIHandle⚠

Sets the stored handle for a SPI port.

HAL_SetSPIMode⚠

Sets the SPI Mode.

HAL_SetSPISpeed⚠

Sets the clock speed for the SPI bus.

HAL_SetSerialBaudRate⚠

Sets the baud rate of a serial port.

HAL_SetSerialDataBits⚠

Sets the number of data bits on a serial port.

HAL_SetSerialFlowControl⚠

Sets the flow control mode of a serial port.

HAL_SetSerialParity⚠

Sets the number of parity bits on a serial port.

HAL_SetSerialReadBufferSize⚠

Sets the size of the read buffer.

HAL_SetSerialStopBits⚠

Sets the number of stop bits on a serial port.

HAL_SetSerialTimeout⚠

Sets the minimum serial read timeout of a port.

HAL_SetSerialWriteBufferSize⚠

Sets the size of the write buffer.

HAL_SetSerialWriteMode⚠

Sets the write mode on a serial port.

HAL_SetSimValue⚠

Sets a simulated value.

HAL_SetThreadPriority⚠

Sets the thread priority for the specified thread.

HAL_SetUserRailEnabled3V3⚠

Enables or disables the 3V3 rail.

HAL_SetUserRailEnabled5V⚠

Enables or disables the 5V rail.

HAL_SetUserRailEnabled6V⚠

Enables or disables the 6V rail.

HAL_SetupAnalogGyro⚠

Sets up an analog gyro with the proper offsets and settings for the KOP analog gyro.

HAL_Shutdown⚠

Call this to shut down HAL.

HAL_SimPeriodicAfter⚠

Calls registered SimPeriodic “after” callbacks (only in simulation mode). This should be called after user code periodic simulation functions.

HAL_SimPeriodicBefore⚠

Calls registered SimPeriodic “before” callbacks (only in simulation mode). This should be called prior to user code periodic simulation functions.

HAL_StartSPIAutoRate⚠

Sets the period for automatic SPI accumulation.

HAL_StartSPIAutoTrigger⚠

Starts the auto SPI accumulator on a specific trigger.

HAL_StopCANPacketRepeating⚠

Stops a repeating packet with a specific ID.

HAL_StopNotifier⚠

Stops a notifier from running.

HAL_StopSPIAuto⚠

Stops an automatic SPI accumulation.

HAL_TransactionI2C⚠

Generic I2C read/write transaction.

HAL_TransactionSPI⚠

Performs an SPI send/receive transaction.

HAL_UpdateNotifierAlarm⚠

Updates the trigger time for a notifier.

HAL_WaitForInterrupt⚠

Waits for the defined interrupt to occur.

HAL_WaitForMultipleInterrupts⚠

Waits for any interrupt covered by the mask to occur.

HAL_WaitForNotifierAlarm⚠

Waits for the next alarm for the specific notifier.

HAL_WriteCANPacket⚠

Writes a packet to the CAN device with a specific ID.

HAL_WriteCANPacketRepeating⚠

Writes a repeating packet to the CAN device with a specific ID.

HAL_WriteCANRTRFrame⚠

Writes an RTR frame of the specified length to the CAN device with the specific ID.

HAL_WriteI2C⚠

Executes a write transaction with the device.

HAL_WriteSPI⚠

Executes a write transaction with the device.

HAL_WriteSerial⚠

Writes data to the serial port.

WPI_CreateEvent⚠

Creates an event. Events have binary state (signaled or not signaled) and may be either automatically reset or manually reset. Automatic-reset events go to non-signaled state when a WaitForObject is woken up by the event; manual-reset events require ResetEvent() to be called to set the event to non-signaled state; if ResetEvent() is not called, any waiter on that event will immediately wake when called.

WPI_CreateSemaphore⚠

Creates a semaphore. Semaphores keep an internal counter. Releasing the semaphore increases the count. A semaphore with a non-zero count is considered signaled. When a waiter wakes up it atomically decrements the count by 1. This is generally useful in a single-supplier, multiple-consumer scenario.

WPI_CreateSignalObject⚠

Sets up signaling for an arbitrary handle. With this function, any handle can operate like an event handle.

WPI_DestroyEvent⚠

Destroys an event. Destruction wakes up any waiters.

WPI_DestroySemaphore⚠

Destroys a semaphore. Destruction wakes up any waiters.

WPI_DestroySignalObject⚠

Cleans up signaling for a handle. Destruction wakes up any waiters.

WPI_ReleaseSemaphore⚠

Releases N counts of a semaphore.

WPI_ResetEvent⚠

Sets an event to non-signaled state.

WPI_ResetSignalObject⚠

Sets a handle to non-signaled state.

WPI_SetEvent⚠

Sets an event to signaled state.

WPI_SetSignalObject⚠

Sets a handle to signaled state.

WPI_WaitForObject⚠

Waits for an handle to be signaled.

WPI_WaitForObjectTimeout⚠

Waits for an handle to be signaled, with timeout.

WPI_WaitForObjects⚠

Waits for one or more handles to be signaled.

WPI_WaitForObjectsTimeout⚠

Waits for one or more handles to be signaled, with timeout.

Type Aliases

HAL_AccelerometerRange

The acceptable accelerometer ranges.

HAL_AddressableLEDHandle

HAL_AllianceStationID

HAL_AnalogInputHandle

HAL_AnalogOutputHandle

HAL_AnalogTriggerHandle

HAL_AnalogTriggerType

The type of analog trigger to trigger on.

HAL_Bool

HAL_CANDeviceType

The CAN device type.

HAL_CANHandle

HAL_CANManufacturer

The CAN manufacturer ID.

HAL_CTREPCMHandle

HAL_CompressorHandle

HAL_CounterHandle

HAL_Counter_Mode

The counter mode.

HAL_DMAHandle

HAL_DigitalHandle

HAL_DigitalPWMHandle

HAL_DutyCycleHandle

HAL_EncoderEncodingType

The encoding scaling of the encoder.

HAL_EncoderHandle

HAL_EncoderIndexingType

The type of index pulse for the encoder.

HAL_FPGAEncoderHandle

HAL_GyroHandle

HAL_Handle

HAL_I2CPort

@defgroup hal_i2c I2C Functions @ingroup hal_capi @{

HAL_InterruptHandle

HAL_MatchType

HAL_NotifierHandle

HAL_PDPHandle

HAL_PortHandle

HAL_PowerDistributionHandle

HAL_REVPDHHandle

HAL_REVPHHandle

HAL_RadioLEDState

HAL_RelayHandle

HAL_RuntimeType

@defgroup hal_capi WPILib HAL API Hardware Abstraction Layer (HAL) to hardware or simulator @{

HAL_SPIMode

HAL_SPIPort

@defgroup hal_spi SPI Functions @ingroup hal_capi @{

HAL_SerialPort

@defgroup hal_serialport Serial Port Functions @ingroup hal_capi @{

HAL_SerialPortHandle

HAL_SimDeviceHandle

HAL_SimValueDirection

Direction of a simulated value (from the perspective of user code).

HAL_SimValueHandle

HAL_SolenoidHandle

HAL_Type

HAL data types.

NativeThreadHandle

@defgroup hal_threads Threads Functions @ingroup hal_capi @{

WPI_EventHandle

An event handle.

WPI_Handle

Generic handle for all WPI handle-based interfaces.

WPI_SemaphoreHandle

A semaphore handle.

Unions

HAL_Value__bindgen_ty_1